Remove drawing_support.py
* Move get_points_for_thick_line to draw_commands.py where it belongs

* Remove drawing_support.py module

* Remove drawing_support.py file and quick index script reference
